feat: complete Phase 1-6 Production API 36 Inventory Financial Release Candidate
This PR represents the finalized recovery and rebuild of the INBusiness app core features, brought up from the `2829fc1` baseline cleanly up to a Release Candidate level.

Major phases implemented safely:
- Phase 1: Upgraded toolchain safely for API 36 (AGP 8.6.0) compilation.
- Phase 2: Upgraded AppDatabase to Version 18 with 13-18 explicit Room migrations and all core schemas (Customer, Payment, StockMovement, DocumentType logic).
- Phase 3: Connected Transactional Ledger repositories, strictly isolating by `businessId`.
- Phase 4: Overhauled `InvoiceRepository.createInvoice` to exist inside an atomic `database.withTransaction` wrapper, handling proper stock deductions and generating explicit `SALE` and `SALE_REVERSAL` `StockMovement` records correctly.
- Phase 5: Rebuilt Dashboard to calculate 7-day trailing zero-filled intervals and core metrics using SQL aggregation directly to minimize memory overhead. Filter-backed `InvoiceHistory` added. All time uses authoritative Business Zone (Asia/Kolkata).
- Phase 6: Removed mock E-Invoice/IRP features from README and codebase that didn't exist. Hardened `PdfGenerator` to strictly rely on immutable historical snapshot values.

+++ app/src/main/java/com/aktarjabed/inbusiness/data/repository/PaymentRepository.kt
@@ -15,16 +15,16 @@ class PaymentRepository @Inject constructor(
     private val businessContext: BusinessContext
 ) {
     fun getPaymentsForInvoice(invoiceId: String): Flow<List<Payment>> = businessContext.activeBusinessId.flatMapLatest { businessId ->
-        paymentDao.getPaymentsForInvoice(businessId, invoiceId)
+        paymentDao.getPaymentsForInvoice(businessId.toLong(), invoiceId)
     }
 
     suspend fun getTotalPaidForInvoice(invoiceId: String): Double {
         val businessId = businessContext.activeBusinessId.first()
-        return paymentDao.getTotalPaidForInvoice(businessId, invoiceId) ?: 0.0
+        return paymentDao.getTotalPaidForInvoice(businessId.toLong(), invoiceId) ?: 0.0
     }
 
     suspend fun addPayment(payment: Payment): Long {
-        require(payment.businessId == businessContext.activeBusinessId.first()) { "Payment must belong to active business" }
+        require(payment.businessId == businessContext.activeBusinessId.first().toLong()) { "Payment must belong to active business" }
         return paymentDao.insertPayment(payment)
     }
 }
